refactor: refactor ccw/ui/doctor_ui.go - large ui file (8...
- Code refactoring

Resolves #249
